Desafio 1963

+++ Python 3/1963.py
@@ -1,11 +1,11 @@
 #Nome: O Filme
-#Resultado: Runtime error
-#Data: 29/07/18 21:42:02
+#Resultado: Accepted
+#Data: 29/07/18 21:42:36
 #Linguagem: Python 3
 entradas = input().split()
 
-dif = int(entradas[1]) - int(entradas[0])
+dif = float(entradas[1]) - float(entradas[0])
 
-percent = dif / int(entradas[0])
+percent = dif / float(entradas[0])
 percent *= 100
 print("{:.2f}%".format(percent))
